+++ ssh_proxy_server/forwarders/ssh.py
@@ -208,89 +208,3 @@ class NoShellForwarder(SSHForwarder):
                     self.session.transport.close()
                     break
             time.sleep(0.5)
-
-
-class SSHInjectableForwarder(SSHForwarder):
-
-    @classmethod
-    def parser_arguments(cls):
-        cls.PARSER.add_argument(
-            '--ssh-injector-net',
-            dest='ssh_injector_net',
-            default='127.0.0.1',
-            help='local address/interface where injector sessions are served'
-        )
-        cls.PARSER.add_argument(
-            '--ssh-injector-port',
-            dest='ssh_injector_port',
-            default=0,
-            type=int,
-            help='local port where injector sessions are served'
-        )
-
-    def __init__(self, session):
-        super(SSHInjectableForwarder, self).__init__(session)
-        self.injector_ip = self.args.ssh_injector_net
-        self.injector_port = self.args.ssh_injector_port
-        self.injector_sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        while True:
-            if self.injector_sock.connect_ex((self.injector_ip, self.injector_port)) == 0:
-                self.injector_port += 1
-            else:
-                self.injector_sock.bind((self.injector_ip, self.injector_port))
-                break
-        self.injector_sock.listen(5)
-
-        self.queue = queue.Queue()
-        self.sender = None
-        self.injector_shells = []
-        thread = threading.Thread(target=self.injector_connect)
-        thread.start()
-        self.conn_thread = thread
-        logging.info("creating ssh injector shell %s, connect with telnet", self.injector_sock.getsockname())
-
-    def injector_connect(self):
-        try:
-            while self.session.running:
-                readable = select.select([self.injector_sock], [], [])[0]
-                if len(readable) == 1 and readable[0] is self.injector_sock:
-                    client, addr = self.injector_sock.accept()
-                    logging.info("injector shell opened from %s", str(addr))
-                    injector_shell = threading.Thread(target=self.injector_session, args=(client,))
-                    injector_shell.start()
-                    self.injector_shells.append(injector_shell)
-        except Exception as e:
-            logging.warning("injector shell suffered an unexpected error")
-            logging.exception(e)
-            self.close_session(self.channel)
-
-    def injector_session(self, client_sock):    # Exception handling
-        with client_sock as sock:
-            while True:
-                data = sock.recv(self.BUF_LEN)
-                self.queue.put((data, sock))
-                time.sleep(0.3)
-
-    def forward_stdout(self):
-        if self.server_channel.recv_ready():
-            buf = self.server_channel.recv(self.BUF_LEN)
-            if self.sender:
-                self.sender.sendall(buf.rstrip())
-                self.sender = None
-                return
-            self.session.ssh_channel.sendall(buf)
-
-    def forward_extra(self):
-        if not self.server_channel.recv_ready() and not self.session.ssh_channel.recv_ready() and not self.queue.empty():
-            msg, sender = self.queue.get()
-            self.server_channel.sendall(msg)
-            self.sender = sender
-            self.queue.task_done()
-
-    def close_session(self, channel):
-        super().close_session(channel)
-        logging.info("closing injector shell %s", self.injector_sock.getsockname())
-        self.injector_sock.close()
-        self.conn_thread.join()
-        for shell in self.injector_shells:
-            shell.join()
